move proposeArgs from test to utils
