+++ contracts/test/helpers.js
@@ -283,26 +283,6 @@ const getAssetAddresses = async (deployments) => {
   }
 };
 
-async function governorArgs({ contract, value = 0, signature, args=[]}) {
-  const method = signature.split("(")[0];
-  const tx = await contract.populateTransaction[method](...args);
-  const data = "0x" + tx.data.slice(10) ;
-  return [tx.to, value, signature, data];
-}
-
-
-async function proposeArgs(governorArgsArray) {
-  const targets=[], values=[], sigs=[], datas=[];
-  for (const g of governorArgsArray) {
-    const [t, v, s, d] = await governorArgs(g);
-    targets.push(t);
-    values.push(v);
-    sigs.push(s);
-    datas.push(d);
-  }
-  return [targets, values, sigs, datas];
-}
-
 module.exports = {
   ousdUnits,
   usdtUnits,
@@ -326,6 +306,4 @@ module.exports = {
   setOracleTokenPriceUsdMinMax,
   getOracleAddresses,
   getAssetAddresses,
-  governorArgs,
-  proposeArgs
 };
